Fixes #18: undefined index with 'label' questions

+++ classes/display.php
@@ -81,7 +81,9 @@ class display extends \mod_feedback_responses_table {
         foreach ($headers as $id => $header) {
             echo \html_writer::start_tag('div', array('class' => 'fitem'));
             echo \html_writer::tag('div', $header, array('class' => 'fitemtitle'));
-            echo \html_writer::tag('div', format_text($formattedrow["val$id"]), array('class' => 'felement fstatic'));
+            if (array_key_exists("val$id", $formattedrow)) {
+                echo \html_writer::tag('div', format_text($formattedrow["val$id"]), array('class' => 'felement fstatic'));
+            }
             echo \html_writer::end_tag('div');
         }
         echo \html_writer::end_tag('div');
